move markdown rendering into internal/markdown package update callers to use this package
