--- cmd/frontend/internal/markdown/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "markdown",
-    srcs = ["markdown.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/markdown",
-    visibility = ["//cmd/frontend:__subpackages__"],
-    deps = [
-        "@com_github_microcosm_cc_bluemonday//:bluemonday",
-        "@com_github_shurcool_github_flavored_markdown//:github_flavored_markdown",
-    ],
-)
